chatapptest/requirements.txt
azure-ai-projects 
azure-identity 
openai 
azure-ai-inference 
azure-search-documents 
azure-ai-evaluation 
azure-monitor-opentelemetry